mkfs/fsck: use PRIu64/PRIx64 to print 64-bit types
Change cpu_to_le64()/cpu_to_le32()/cpu_to_le16() defines to
return determined types (instead of __le64/__le32/__le16) and
use PRIu64/PRIx64 to print 64-bit types.

Fixes:

  mkfs.c:60:14: error: format ‘%llu’ expects argument of type ‘long long unsigned int’, but argument 2 has type ‘long unsigned int’ [-Werror=format=]
    exfat_debug("Volume Length(sectors) : %llu\n",
                ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

  fsck.c:484:13: error: format ‘%llu’ expects argument of type ‘long long unsigned int’, but argument 2 has type ‘long unsigned int’ [-Werror=format=]
     exfat_err("too large sector count: %llu\n, expected: %llu\n",
               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

  fsck.c:851:13: error: format ‘%llu’ expects argument of type ‘long long unsigned int’, but argument 2 has type ‘long unsigned int’ [-Werror=format=]
     exfat_err("valid size %llu greater than size %llu: %s\n",
               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

  fsck.c:930:14: error: format ‘%llx’ expects argument of type ‘long long unsigned int’, but argument 3 has type ‘long unsigned int’ [-Werror=format=]
     exfat_debug("start cluster %#x, size %#llx\n",
                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

  fsck.c:938:13: error: format ‘%llx’ expects argument of type ‘long long unsigned int’, but argument 2 has type ‘long unsigned int’ [-Werror=format=]
     exfat_err("invalid size of allocation bitmap. 0x%llx\n",
               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

  fsck.c:992:13: error: format ‘%llx’ expects argument of type ‘long long unsigned int’, but argument 2 has type ‘long unsigned int’ [-Werror=format=]
     exfat_err("invalid size of upcase table. 0x%llx\n",
               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ include/exfat_ondisk.h
@@ -26,9 +26,9 @@
 #define cpu_to_le64(x)	(x)
 #endif
 
-#define le64_to_cpu(x)  cpu_to_le64(x)
-#define le32_to_cpu(x)  cpu_to_le32(x)
-#define le16_to_cpu(x)  cpu_to_le16(x)
+#define le64_to_cpu(x)  ((uint64_t)cpu_to_le64(x))
+#define le32_to_cpu(x)  ((uint32_t)cpu_to_le32(x))
+#define le16_to_cpu(x)  ((uint16_t)cpu_to_le16(x))
 
 #define PBR_SIGNATURE		0xAA55
